fix(ui): Split reducers into individual files (AEROGEAR-8922)
