+++ ui/src/reducers/header.js
@@ -1,0 +1,10 @@
+const initialState = {
+  currentUser: 'currentUser'
+};
+
+export default (state = initialState, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
